==> Step 3a Expand our Data -  Get Deprivation Data.ipynb <==
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.common.by import By
from selenium.common.exceptions import NoSuchElementException
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import os

import pandas as pd

# open the initial webpage and logon

chrome_driver_path = 'C:\\chromedriver.exe'
browser = webdriver.Chrome(service=Service(chrome_driver_path))
browser.get("https://imd-by-geo.opendatacommunities.org/imd/2019/area")


# get all the authorities in the district grid

link = browser.find_element(By.XPATH, '//*[@id="district-grid"]')
URLS = link.find_elements(By.TAG_NAME,'li')

## developers have split the data up into 4 divs so we need to loop through each div to get the data, skipping any files that give an error

Div = 1

while Div <= 4:

    row = 1
    while row <= len(URLS):

        # always go back to the start screen   
        browser.get("https://imd-by-geo.opendatacommunities.org/imd/2019/area")

        #need to capture no element error
        try:
            authority= browser.find_element(By.XPATH, '//*[@id="district-grid"]/div['+str(Div)+']/div/ul/li['+str(row)+']/a')  
            authority_name = authority.text
            print("Processing ", Div, ":", row, ":", authority_name)
    
            click = authority.get_attribute("href")
            browser.get(click)

            Step2 = browser.find_element(By.XPATH, '//*[@id="step_two"]/form/input[1]')
            Step2.submit()
            
            Step3 = browser.find_element(By.XPATH, '//*[@id="step_three"]/form/input[1]')
            Step3.submit()

        except NoSuchElementException:
            print("Skipping ", Div, ":", row, ":")
        row += 1
   
    Div += 1
    

## Now read all of these files into a dataframe and then save it as a CSV

import pandas as pd
import glob
import os

path = r'C:\Users\nigel\OneDrive\Documents\GitHub\ig\FoodRating\Deprivation Files' # use your path
all_files = glob.glob(os.path.join(path , "*.csv"))

working_list = []

for filename in all_files:
    #print (filename)
    df = pd.read_csv(filename, index_col=None, header=0)
    working_list.append(df)

frame = pd.concat(working_list, axis=0, ignore_index=True)
frame = frame.drop(["Postcode Status","LSOA Code", "LSOA Name", "Index of Multiple Deprivation Rank"], axis=1 )

# Make the PostCode_S field



for index, row in frame.iterrows():
    PostCode_S = row['Postcode'].replace(" ","")
    frame.loc[index, 'PostCode_S'] = PostCode_S


dummy = frame.to_csv("Deprivation Data", index = False)


frame
