==> Step 3b Expand our Data.ipynb <==
## Build a lookup file using the postcode translation table as basis

## first, lets standadise the postocde field by removing spaces

import pandas as pd

pcd = pd.read_csv("PCD.csv")
for index, row in pcd.iterrows():
    PostCode_S = row['pcd'].replace(" ","")
    pcd.loc[index, 'PostCode_S'] = PostCode_S
    pcd_sect_S = row['pcd_sect'].replace(" ","")
    pcd.loc[index, 'pcd_sect_S'] = pcd_sect_S


## get a list of all the postcode sectors

## merge with our data file

FR = pd.read_csv("FoodRating.csv")
# Only select records with rathing 0-5

FR = FR[(FR.RatingValue).isin(["0", "1", "2", "3", "4", "5"])]
FR = FR[(FR.PostCode) != "-"]
FR["RatingValue"]= FR["RatingValue"].astype('int')

len(FR)

pcd_FR = pd.merge(pcd, FR, how = "inner", on =['PostCode_S', 'PostCode_S'])

len(pcd_FR)

#Calculate the mean by PostCode Sector

pcd_FR_mean_sect = pcd_FR.groupby(['pcd_sect'], sort=False)['RatingValue'].mean().reset_index()

pcd_FR_mean_sect2 = pcd_FR_mean_sect.rename(columns={'RatingValue': 'RVM_Sector'})

pcd_working = pd.merge(pcd, pcd_FR_mean_sect2, how = "left", on =['pcd_sect', 'pcd_sect'])

len(pcd_working)

#Calculate the mean by PostCode Area

pcd_FR_mean_area = pcd_FR.groupby(['pcd_area'], sort=False)['RatingValue'].mean().reset_index()

pcd_FR_mean_area = pcd_FR_mean_area.rename(columns={'RatingValue': 'RVM_Area'})

pcd_working = pd.merge(pcd_working, pcd_FR_mean_area, how = "left", on =['pcd_area', 'pcd_area'])

#Calculate the mean by PostCode District

pcd_FR_mean_dist = pcd_FR.groupby(['pcd_dist'], sort=False)['RatingValue'].mean().reset_index()

pcd_FR_mean_dist = pcd_FR_mean_dist.rename(columns={'RatingValue': 'RVM_Dist'})

pcd_working = pd.merge(pcd_working, pcd_FR_mean_dist, how = "left", on =['pcd_dist', 'pcd_dist'])

pcd_ethnic = pd.read_csv("PCD Ethnic.csv")


pcd_ethnic

for index, row in pcd_ethnic.iterrows():
   
    pcd_sect_S = row['pcd_sect'].replace(" ","")
    pcd_ethnic.loc[index, 'pcd_sect_S'] = pcd_sect_S
    


pcd_ethnic = pcd_ethnic.drop(["pcd_sect"], axis=1)

pcd_working = pd.merge(pcd_working, pcd_ethnic, how = "left", on =['pcd_sect_S', 'pcd_sect_S'])

len(pcd_working)

pcd_working

pcd_deprivation = pd.read_csv("Deprivation Data.csv")

pcd_deprivation

pcd_working = pd.merge(pcd_working, pcd_deprivation, how = "left", on =['PostCode_S', 'PostCode_S'])


pcd_working = pcd_working.drop(["pcd", "pcd2", "pcds", "pcd_dist", "pcd_area", "oa21"], axis =1 )

pcd_working

pcd_working = pcd_working.rename(columns={'Index of Multiple Deprivation Decile': 'Deprivation'})

pcd_working

data = pcd_working.to_csv("pcd_FR_mean_ethnic.csv", index=False)
