==> Step 3c Expand our Data.ipynb <==
## Results

## Load and understand our myData Data

from sklearn import datasets
import pandas as pd
import numpy as np


FR = pd.read_csv('FoodRating.csv')

FR.dtypes

PCD = pd.read_csv('pcd_FR_mean_ethnic.csv')

PCD

## Lets use the Food Ratings File and lookup the Ratings, Etnicity and Deprivation data 

myData = pd.merge(FR, PCD, how = "left", on =['PostCode_S', 'PostCode_S'])


myData.keys()

myData.shape

myData.head()

## Pre-process our data

import pandas as pd


df_merge3 = myData
df_merge3["RatingValue"]= df_merge3["RatingValue"].astype('int')
df_merge3["Business Type Index"]= df_merge3["Business Type Index"].astype('category')
df_merge4 = df_merge3.dropna()
print ("After getting rid of the NA's we are left with:", len(df_merge4))





df_merge4.dtypes

dummy = df_merge4.to_csv("Data.csv", index = False)
