==> Step 5 and 6 Make a Tool and Run the Model against all our data.ipynb <==
## Choose our model

import pickle
model = pickle.load(open('clf_LGBM.pickle', 'rb'))

## Localised Test

import pandas as pd
PCD = pd.read_csv("pcd_FR_mean_ethnic.csv")

Data_Set = 1

# we need to load the entire postcode file as we will use it as a lookup on the PostCode to the attributes to feed into our model..

postcodes = []
#postcodes.append('SS141JB')
postcodes.append('EN87BW')
postcodes.append('CB11PA')
postcodes.append('CM11PH')
postcodes.append('CM20BX')
postcodes.append('LU14LH')
postcodes.append("BB97AG")
postcodes.append("B81HN")
postcodes.append("BB95HY")
postcodes.append("L339TW")
postcodes.append("BB16EA")
postcodes.append("BS494HS")

test_BTI = []
test_RVM_sect = []
test_RVM_area = []
test_RVM_dist = []
test_White = []
test_Mixed = []
test_Asian = []
test_BAC = []
test_Other = []
test_Deprivation = []
test_GeoCode_Lat = []
test_GeoCode_Long = []

for postcode in postcodes:
    row = PCD.loc[PCD['PostCode_S'] == postcode]
   # print (row)
    row2 = row.to_numpy()
    test_BTI.append ("1")
    test_RVM_sect.append(row2[:,3])
    test_RVM_area.append(row2[:,4])
    test_RVM_dist.append(row2[:,5])
    test_White.append(row2[:,6])
    test_Mixed.append(row2[:,7])
    test_Asian.append(row2[:,8])
    test_BAC.append(row2[:,9])
    test_Other.append(row2[:, 10])
    test_Deprivation.append(row2[:,12])

if Data_Set ==1:
    test_data = pd.DataFrame(list(zip(test_BTI, test_RVM_sect, test_RVM_area, test_RVM_dist, test_White, test_Mixed, test_Asian, test_BAC, test_Other, test_Deprivation)), columns = ['Business Type Index', 'RVM_Sector', 'RVM_Area', 'RVM_Dist', 'White', 'Mixed', 'Asian', 'BAC', 'Other', 'Deprivation'])
elif Data_Set == 2:
    test_data = pd.DataFrame(list(zip(test_White, test_Mixed, test_Asian, test_BAC, test_Other, test_Deprivation)), columns = [ 'White', 'Mixed', 'Asian', 'BAC', 'Other', 'Deprivation'])
elif Data_Set == 3:
    test_data = pd.DataFrame(list(zip(test_GeoCode_Lat, test_GeoCode_Long, test_White, test_Mixed, test_Asian, test_BAC, test_Other, test_Deprivation)), columns = [ 'White', 'Mixed', 'Asian', 'BAC', 'Other', 'Deprivation'])
    
test_data.dropna(inplace=True)

test_data["RVM_Sector"]= test_data["RVM_Sector"].astype('float')
test_data["RVM_Area"]= test_data["RVM_Area"].astype('float')
test_data["RVM_Dist"]= test_data["RVM_Dist"].astype('float')
test_data["White"]= test_data["White"].astype('float')
test_data["Mixed"]= test_data["Mixed"].astype('float')
test_data["Asian"]= test_data["Asian"].astype('float')
test_data["BAC"]= test_data["BAC"].astype('float')
test_data["Other"]= test_data["Other"].astype('float')
test_data["Deprivation"]= test_data["Deprivation"].astype('float')

test_data["Business Type Index"]= test_data["Business Type Index"].astype('category')
#test_data = test_data.astype({col:value})


test_data.dtypes

y_pred = model.predict(test_data)

y_pred

## Ultimate test


# load our data set and create 
Data = pd.read_csv("Data.csv")
Data["RVM_Sector"]= Data["RVM_Sector"].astype('float')
Data["RVM_Area"]= Data["RVM_Area"].astype('float')
Data["RVM_Dist"]= Data["RVM_Dist"].astype('float')
Data["White"]= Data["White"].astype('float')
Data["Mixed"]= Data["Mixed"].astype('float')
Data["Asian"]= Data["Asian"].astype('float')
Data["BAC"]= Data["BAC"].astype('float')
Data["Other"]= Data["Other"].astype('float')
Data["Deprivation"]= Data["Deprivation"].astype('float')
Data["Business Type Index"]= Data["Business Type Index"].astype('category')

X = Data.drop(['Business Name', 'Business Type', 'RatingValue', 'PostCode', 'RatingDate', 'LocalAuthorityName' ,'LocalAuthorityCode', 
                            'Hygiene', 'Structural', 'Confidence In Management', 'PostCode_S', 'PostCode', "pcd_sect","pcd_sect_S","Geocode_Lat", "Geocode_Long", "Postcode"] , axis=1)

y_pred = model.predict(X)



answer = pd.DataFrame(y_pred, columns = ["Predicted Result"])

result = pd.concat([Data, answer], axis = 1)

result

summary = pd.pivot_table(result, index = ['LocalAuthorityName'], values = ["Predicted Result", "RatingValue"])

dummy = summary.to_csv("summary.csv")
